+++ lib/rubocop/cop/style/variable_name.rb
@@ -18,25 +18,15 @@ module RuboCop
           check_name(node, name, node.loc.name)
         end
 
-        # TODO: Why is this checking invocations of setter rather than
-        #   definitions? Also, this is not variable.
-        def on_send(node)
-          return unless setter_call_on_self?(node)
-          _receiver, method_name, = *node
-          attribute_name = method_name.to_s.sub(/=$/, '').to_sym
-          check_name(node, attribute_name, node.loc.selector)
+        def on_cvasgn(node)
+          name, = *node
+          check_name(node, name, node.loc.name)
         end
 
         private
 
         def message(style)
-          format('Use %s for variables.', style)
-        end
-
-        def setter_call_on_self?(send_node)
-          receiver, method_name, = *send_node
-          return false unless receiver && receiver.type == :self
-          method_name.to_s.end_with?('=')
+          format('Use %s for variable names.', style)
         end
       end
     end